Prevent overwriting user-defined supported completion function
To integrate with vim-lsp this plugin overwrites the default function for
getting LSP capabilities to check for snippet capabilities. However when doing
this, it first calls the default capabilities function before adding its own
keys into the request. This caused any user-defined functions that overwrite the
default to be ignored.  Instead, this plugin should use whatever is set as the
global capabilities function value (which falls back to the default) to ensure
that any user-defined functions aren't ignored